Add message and parameter description to tooltips
Displays message or parameter description when hovering over identifier usages. Requires "Color identifiers" and "Replace Visual Studio tooltip" enabled (or Enhanced Tooltip installed). It doesn't appear in code completion for method invocation or parameter info (because this is powered from "parameter info" that requires XML docs, and doesn't use description providers). Surprisingly, it will show in the tooltip when a parameter is listed in code completion, because that uses the description provider.