Update .gitignore for the IntelliJ project
Sets up project:
* Gradle auto-import
* Issue navigation in commit comments
* Sets up task server (but no details)

Note that gradle.xml is frequently added to .gitignore. This is because
it can contain per-user path information. We're using the Gradle
wrapper, so there's no path information. Adding it means we don't have
to set up the Gradle project on a fresh clone.

Also note that we've ignored the .iml files. These are auto-generated
for a Gradle project, so there's no need to check them in, and they will
change whenever we update the build.gradle file.