Attempt to workaround build issue
I think this is a .net core 2 nuget issue. Looks like the multiple sources are passed through as a semi-colon delimited concatenated string, rather than individually. This causes the validation to fail, as it's treated as a local path, and semi-colon is invalid. Swapping the order passes validation, and I *think* the string is correctly split afterwards.

https://github.com/NuGet/Home/issues/5817