Add temp workaround for failing AppVeyor builds
After updating to Gradle 4.10, AppVeyor started treating warnings output
by rdgen as errors, stopping the build. I think that post 4.7 (possibly
in 4.8) Gradle changed the behaviour of the standard output streams. My
theory is that AppVeyor's powershell implementation sees the warnings on
the error stream and the ErrorActionPreference variables means to treat
warnings as errors, and boom. This change will temporarily change the
ErrorActionPreference variable to hopefully allow rdgen to grumble
without killing the build