Correctly override existing event functions
If an event function is defined in a base class, do not offer to generate it via code completion or the generate dialog. If an event function is defined as a virtual function in a base class, correctly generate the override from the generate dialog, and rely on the standard override code completion.

Also groups event functions by declaring type, or implementing type for virtual functions.

Fixes RIDER-17104, fixes #603, fixes #879, fixes #810