Handle namespace suggestions for external packages
Works for packages that live at the root of a Unity solution, but doesn't work for packages external to the solution. External files are added to a project with a full path, and Rider will find the common ancestor folder and create a linked folder. This isn't included in namespace suggestions, and means we don't get a full folder structure for correct suggestions.

The solution would be to add Link attributes to describe the correct location relative to the containing package root.