Disable zone check inspections
They don't work properly with plugins - the checks are designed to work with zone definitions defined in source code, not consumed from references.
There is also a module consistency check to ensure that zones for namespaces in different projects are consistent. This assumes all modules are related, and doesn't work with multiple modules that have no relationship. I.e. it tries to make the Rider and ReSharper modules consistent, which can't work because one will require VS zones, and the other will require Rider/ReSharperHost zones.