Update API information to 2019.4.15f1
This includes EditorWindow.CreateGUI, which should have been added when updating to 2019.4.4f1. I have no idea why it didn't wasn't. The API is also included in 2020.1 and 2020.2. The api.xml file is only up to date with beta and alpha versions, so it's possible that the new API was missing at that point.