Prevent break on unhandled ExitGUIException
The behaviour for unhandled exceptions has changed in the Unity 2021.2 beta, thanks to a Mono upgrade. ExitGUIException is frequently thrown in the GUI, but handled in native code. Mono now reports this as an unhandled exception, which is very annoying and makes it hard to use the editor while debugging.