Use 9999 as build number from the command line
The 9998 build number was intended to show that the plugin had been
built from the IDE, or from msbuild, while 9999 would indicate that it
was built from gradle. This could possibly help diagnose build errors.
However, due to strong naming, it causes unnecessary warnings and
makes it more difficult to update the ReSharper plugin once installed.