Rename platform types to include _ALL postfix
As we now support aarch64 for all OSs, we have extended platform types
so that each test could be muted or executed on any combination of OS
and architecture. Not to cause any confusion, WINDOWS/MAC_OS/LINUX
platforms were renamed to have *_ALL postfix which indicates that given
test will be executed on all architectures of the given OS.

In future we might want to extend it to specify architectures, so that
test can be executed on all OSs for the given arch. This will be
evaluated by the usages.