fix RIDER-91186 Add versioning for IDeferredCache's
Add DeferredCacheBase.Version and ClearOnLoad

For ordinary ICacheWithVersion/SimpleICache, the corresponding logic is partially in SimpleICache and partially in PsiCacheRegistry.
In DeferredCacheBase, timestamps are stored inside cache (and updated on Merge), so everything can be handled inside the class. In particular, we:
(1) clear myTimestamps if ClearOnLoad == true
(2) don't need to check ClearOnLoad in DeferredCacheBase.UpToDate(), in contrast to SimpleICache.UpToDate()