Move external sources settings override to Rider only
The overridden settings will ensure that the user will navigate to source for compiled symbols, which is important for Rider because we allow the user to browse the (external) source of packages through the Unity Explorer, so we should also navigate to the same sources. ReSharper does not have a Unity Explorer, so we should leave the choice up to the user.

Fixes RIDER-90596

(cherry picked from commit 1d7fefebb0e0705ab55ef0e04e631db0e201fbb5)