Remove the override of type conversion hint settings
The settings were overridden to ensure that lossy implicit conversions between Unity's Vector2, Vector3 and Vector4 were visible, but this stops a user from choosing an alternative default. The product's default is to always show type conversion hints as icons, so unless the user changes anything, this is always true. If the user does change the global setting, we should respect it.

Fixes RIDER-90596