[Workspace Model] Remove `entityStorage` from the `WorkspaceModel` API
To get a version of the storage, `currentSnapshot` should be used.

`WorkspaceModelImpl` and `VersionedEntityStorage` are marked as internal. `WorspaceModelImpl` should has an internal modifier, but this is currently not possible because Rider overrides it. `VersionedEntityStorage` may be removed in future at all.

There are some cases when the version of the storage is accessed. Probably we can create a special function to return an entity storage with its version.

This is a part of IJPL-345

GitOrigin-RevId: cc3d6c47099c91c54d571d7ca13c5ef9d84265a9